基本的にinterfaceのあるpackageに対応抽象クラスをおいといて、そのクラスで参照する系のデータは 保持させることにする。

+++ myLib.MIT/myLib.container.flv/src/main/java/com/ttProject/container/flv/FlvTag.java
@@ -40,7 +40,7 @@ public abstract class FlvTag implements IContainer {
 	 * flvは全体サイズとは別に２つのサイズがあるので注意
 	 */
 	@Override
-	public long getSize() {
+	public int getSize() {
 		if(dataSize == null) {
 			return -1;
 		}
